New movement system implementation (1/3)
added a source-like movement system. It work now but i need to repair everything that it broke when i've implemented it.

source of the new system :
https://github.com/Olezen/UnitySourceMovement

also :
made some change to the room geometry, and the school (2/3 will be the optimization of the school's geometry and mesh because HOLY SHIT it's a fucking mess)